refactor(identity): collapse genome::recall::PeerId onto canonical airc PeerId
The third of four identity shapes. `genome::recall::PeerId(Uuid)` was a second
continuum-side re-invention of `airc_core::PeerId` (the federation-peer addressing
type used in forge/*) — same primitive, same name, different module. A federation
peer and a persona are the SAME actor; they must be the same type.

- Delete the `recall::PeerId` newtype + its ts-rs export; recall.rs now imports
  `crate::identity::PeerId` (so the public ResidencyHint::GridPeer /
  RecallScope::Federation fields expose the canonical type).
- Migrate all consumers to `crate::identity::PeerId`: forge/{endpoint,gene_handle,
  grid_custodian}.rs, genome/{recall_scoring,recall_trait}.rs; drop `PeerId` from
  the genome/mod.rs re-export (nobody imported `genome::PeerId`).
- `PeerId::new(uuid)` → `from_uuid(uuid)` at every wrap-known-uuid site.
- Wire fields embedding the id keep their string shape via `#[ts(type=...)]`:
  `string` for scalar (ForgeLocator/GeneLocator::Node.node, GridPeer.peer),
  `Array<string>` for Federation.peers, `Array<[string, TrustClass]>` for
  RecallContext.trust_overrides (PeerId is serde-transparent, no ts-rs derive).
- recall_trait.rs: collapse the prior fully-qualified `crate::identity::PeerId`
  workaround now that the import no longer collides with recall::PeerId.

`cargo check --features metal,accelerate` green. Only the bare-Uuid identity
fields (CallerIdentity, PersonaInstanceInfo) remain — Step 4. Part of #38.
